requirements.txt
# 基础依赖
langchain
langgraph
litellm
python-dotenv
pyyaml
requests
langsmith

# 可选依赖（用于OpenAI模型支持）
# langchain-openai